Fixed major (subtle) error in calc_hessian.
Changed hessian calculation to properly handle boundary
vertices. Now uses curvature formula
	K = Pi - angle sum
for boundary vertices, which results in a symmetric
hessian. Alternative is
	K = 2Pi - 2*(angle sum)
but this is an unsymmetric hessian.

Also, combined angle data in a single entry, so now each
dKi / duj term only occupies a single entry in the sparse matrix.
(much faster multiplication as a result).